+++ tests/world/test_world_integration.py
@@ -12,13 +12,12 @@ def test_bot_move_in_world_context() -> None:
     # arrange
     w = World(10)
     b = Bot(
-        world=w,
         name="b0",
         position_from_sequence=(0, 0),
     )
+    w.add_entity(b)
     b.velocity = Vector2(1, 0)
-
     # act
     w.update()
-
+    # assert
     assert b.position == Vector2(1 / SIMULATION_FPS, 0)